% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{AddPropCI}
\alias{AddPropCI}
\title{Add a Binomial Proportion Confidence Interval to a Data Frame}
\usage{
AddPropCI(data, p = "p", n = "n", lb = "lb", ub = "ub", conf_level = 0.95)
}
\arguments{
\item{data}{A data frame containing the sample proportions and sample sizes
for computing the proportion confidence interval.}

\item{p}{Name of column containing the sample proportions.}

\item{n}{Name of column containing the sample sizes.}

\item{lb}{Name of column to contain lower bound of confidence interval.}

\item{ub}{Name of column to contain upper bound of confidence interval.}

\item{conf_level}{Confidence level for interval.}
}
\value{
The data frame passed into the function with the binomial proportion
  confidence interval in the \code{lb} and \code{ub} columns.
}
\description{
Adds the lower and upper bounds of a binomial proportion confidence interval
to a data frame.
}
