% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{AdjustedEffectFormula}
\alias{AdjustedEffectFormula}
\title{Formulas for Linear Models to Estimate Regression-adjusted Effects}
\usage{
AdjustedEffectFormula(
  y,
  vars = NULL,
  int_vars = NULL,
  treat = "treat",
  model = c("ancova", "did"),
  include_pre = TRUE,
  interact_pre = TRUE,
  paired = FALSE,
  var_equal = TRUE,
  match = "match",
  cluster = NULL,
  panel_time = NULL,
  unit = NULL,
  post_suffix = "_post"
)
}
\arguments{
\item{y}{Character vector containing names of metrics. These are names of the
metrics without the \code{_post} and \code{_pre} suffixes.}

\item{vars}{Character vector containing the names of covariates to include
in the regression model without interacting with the treatment variable.
These do not need to include the pre-period values of the metrics specified
in \code{y} if \code{include_pre = TRUE}.}

\item{int_vars}{Character vector containing the names of covariates to
include in the linear model for interacting with the treatment variable.
These do not need to include the pre-period values of the metrics specified
in \code{y} if both \code{include_pre = TRUE} and
\code{interact_pre = TRUE} hold.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{model}{The type of regression model, whether \code{ancova} for a
unconstrained ANCOVA model or \code{did} for a difference-in-differences
model.}

\item{include_pre}{Whether to include pre-period metrics in the regression.}

\item{interact_pre}{Whether to interact pre-period metrics with the treatment
variable in the regression. Applicable only when \code{include_pre = TRUE}.}

\item{paired}{Whether the data is paired, in which case a mixed-effects
model with a random effect for the matching variable \code{match} is
fitted using \code{lme4::lmer}.}

\item{var_equal}{Whether to assume unequal variances for the metrics across
treatment and control group. Similar to the \code{var.equal} argument in
the \code{stats::t.test} function. Applies only when \code{paired = FALSE}.}

\item{match}{Name for variable for matching paired subjects.}

\item{cluster}{Name of variable for clustering the data. Defaults to NULL,
in which case, a linear model will be fit using \code{stats::lm}. If
\code{cluster} is not NULL, or \code{paired | !var_equal}, a linear
mixed-effects model will be fit using \code{lme4::lmer}.}

\item{panel_time}{Name of time variable for longidudinal or panel data.}

\item{post_suffix}{Suffix for the post-period metric, which will have a
string name equal to \code{paste0(y, post_suffix)}. Defaults to
\code{"_post"}. Pass in the empty string \code{""} to omit the suffix.}
}
\value{
A data frame containing the variable names and model formula for
  fitting regression models to estimate regression-adjusted treatment
  effects.
}
\description{
Create a data frame with variable names and model formula for fitting
regression model to estimate regression-adjusted treatment effects using
\code{AdjustedEffectTable}.
}
