% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_effects.R
\name{AdjustedEffectTable}
\alias{AdjustedEffectTable}
\title{Table of Regression-adjusted Treatment Effects for Pre-post Analysis}
\usage{
AdjustedEffectTable(
  data,
  y,
  vars = NULL,
  int_vars = NULL,
  treat = "treat",
  unit = NULL,
  estimand = c("ate", "att", "atu", "atc"),
  model = c("ancova", "did"),
  include_pre = FALSE,
  interact_pre = FALSE,
  weights = NULL,
  use_svyglm = FALSE,
  paired = FALSE,
  var_equal = TRUE,
  match = "match",
  cluster = NULL,
  panel_time = NULL,
  post_suffix = "_post",
  conf_level = 0.95,
  rel_eff = TRUE,
  hc = c("none", "hc0", "hc1", "hc2", "hc3", "hc4"),
  output_models = FALSE,
  trim_fit = TRUE,
  reff_ci_method = c("fieller", "delta", "delta_log"),
  adjusted_means = TRUE,
  ylab = "metric",
  percent_vars = c("mC", "mT", "diff", "mCa", "mTa", "eff", "se", "pval", "lb", "ub",
    "sd", "smd"),
  format_vars = "stat",
  digits = 1
)
}
\arguments{
\item{data}{A data frame containing metrics for treatment and control
subjects, an indicator variable for treatment assignment and optionally
additional covariates and an integer variable for matching of pairs.}

\item{y}{Character vector containing names of metrics. These are names of the
metrics without the \code{_post} and \code{_pre} suffixes.}

\item{vars}{Character vector containing the names of covariates to include
in the regression model without interacting with the treatment variable.
These do not need to include the pre-period values of the metrics specified
in \code{y} if \code{include_pre = TRUE}.}

\item{int_vars}{Character vector containing the names of covariates to
include in the linear model for interacting with the treatment variable.
These do not need to include the pre-period values of the metrics specified
in \code{y} if both \code{include_pre = TRUE} and
\code{interact_pre = TRUE} hold.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{unit}{Name of identifier variable for each study unit, e.g.,
\code{obscured_gaia_id} for users in the context of Cloud experiments.
Defaults to \code{NULL}. This is needed when each unit has potentially
more than one measurement or observation, e.g., when \code{panel_time} is
specified, in the case of longitudinal or panel data,}

\item{estimand}{The causal estimand to be estimated by the regression
adjustment model, either \code{"ate"} for the average treatment effect on
the entire population (ATE), \code{"att"} for the average treatment effect
on the treated (ATT), or \code{"atu"} or \code{"atc"} for the average
treatment effect on the untreated or control (ATU or ATC).}

\item{model}{The type of regression model, whether \code{ancova} for a
an ANCOVA model, where the response in the post-period metric, or
\code{did} for a difference-in-differences model, where the response is the
difference between the post- and pre-period metrics.}

\item{include_pre}{Whether to include pre-period metrics in the regression.}

\item{interact_pre}{Whether to interact pre-period metrics with the treatment
variable in the regression. Applicable only when \code{include_pre = TRUE}.}

\item{weights}{Name of variable containing inverse-probability-of-treatment
weights (IPTW) for a weighted data sample. We recommend using a
heteroskedasticity-corrected standard errors in this case, e.g.,
\code{hc = "hc0"} instead of \code{hc = "none"}.}

\item{use_svyglm}{Whether to use \code{survey::svyglm} to estimate an IPTW
treatment effect effect with its standard errors. The default is FALSE,
in which case \code{stats::lm} is used.}

\item{paired}{Whether the data is paired, in which case a mixed-effects
model with a random effect for the matching variable \code{match} is
fitted using \code{lme4::lmer}.}

\item{var_equal}{Whether to assume unequal variances for the metrics across
treatment and control groups. Similar to the \code{var.equal} argument in
the \code{stats::t.test} function. Applies only when \code{paired = FALSE}.}

\item{match}{Name for variable for matching paired subjects.}

\item{cluster}{Name of variable for clustering the data. Defaults to NULL,
in which case, a linear model will be fit using \code{stats::lm}. If
\code{cluster} is not NULL, or \code{paired | !var_equal}, a linear
mixed-effects model will be fit using \code{lme4::lmer}.}

\item{panel_time}{Name of time variable for longidudinal or panel data.}

\item{post_suffix}{Suffix for the post-period metric, which will have a
string name equal to \code{paste0(y, post_suffix)}. Defaults to
\code{"_post"}. Pass in the empty string \code{""} to omit the suffix.}

\item{conf_level}{Confidence level for inference.}

\item{rel_eff}{Whether to compute the relative effect and its CI using the
Fieller or delta method.}

\item{hc}{Option for computing heteroskedasticity-corrected SEs using
\code{car::hccm}. Defaults to \code{"none"}. The other values are
\code{"hc0"}, \code{"hc1"}, \code{"hc2"}, \code{"hc3"}, and \code{"hc4"}.
See documentation for \code{car::hccm} for details.}

\item{output_models}{Whether to return the regression model objects together
in addition to the effect table. In this case, a list will be returned
containing the effect table in the first entry and a list of model objects
in the second entry.}

\item{trim_fit}{Whether to apply the \code{TrimLinearModelFit} function to
a model object. Currently works for \code{stats::lm} and
\code{survey::svyglm} objects, i.e., only for the unpaired and
equal-variance cases.}

\item{reff_ci_method}{A string for the relative effect CI method, either
\code{"fieller"} for the Fieller CI, \code{"delta"} for the delta method
without log transformation of the ratio, or \code{"delta_log"} for the
delta method with log transformation of the ratio. See documentation for
the \code{RatioCI} function for details.}

\item{adjusted_means}{Whether to show the model-estimated average potential
outcomes for, respectively, not receiving the treatment, \code{"mCa"}, and
receiving the treatment, \code{"mTa"}.}

\item{ylab}{Label for column containing names of metrics.}

\item{percent_vars}{Names for variables to format as percentages. Note that
this only changes the printing format, not the actual value of the
variables.}

\item{format_vars}{Names for variables to format as a float. Note that this
only changes the printing format, not the actual value of the variables.}

\item{digits}{Number of decimals for table statistics.}
}
\value{
A data frame containing the table of regression-adjusted effects.
}
\description{
Estimate regression-adjusted treatment effects for multiple metrics using
a linear model, and summarize the effects and inference statistics in a
table. The metric is regressed on a treatment indicator variable and
potentially the following:
(1) Covariates not interacted with the treatment variable, specified in the
    \code{vars} argument, and
(2) Covariates interacted with the treatment variable, specified in the
    \code{int_vars} argument
All covariates are centered, i.e., have their means subtracted from their
values, so that:
(1) The intercept coefficient becomes the average control response, and
(2) The coefficient for treatment variable becomes the average treatment
    effect, i.e., the average treatment response minus the average control
    response.
Users can include the pre-period value of the metric in the model, in which
case a pre-post analysis is conducted. We will refer to the pre-period value
as the pre-period metric, and the experiment-period value as the post-period
metric. Whether the pre-period metric is included in the model or not, the
name of the post-period metric must always follow the convention
\code{paste(y, "_post")}, where \code{y} is the vector metric names without
the time suffixes of "_post" and "_pre".
Users can specify two model types through the \code{model} argument:
(1) ANCOVA (\code{ancova}), regression of metric on treatment variable and
    optionally other covariates, including the pre-period metric, or
(2) Difference-in-difference (\code{did}), regression of difference between
    post-period and pre-period metrics on the treatment variable and
    optionally other covariates.
The difference-in-difference model is generally fitted without additional
covariates, but we provide the option for the user to include them as well.
The ANCOVA model without interacted covariates is known as the equal-slopes
ANCOVA model, while that with interacted covariates is known as the
unequal-slopes ANCOVA model. We have Lin's (2013) model when the following
conditions hold:
(1) The covariates interacted with the treatment are the same as those not
    interacted with the treatment, and
(2) All covariates are centered.
That is, Lin's (2013) model is an unequal-slopes ANCOVA model with centered
covariates, all of which appear twice, both as a non-interaction term and as
a treatment-interaction term.
}
\references{
\itemize{
\item Lin, Winston (2013). "Agnostic Notes on Regression Adjustments to
  Experimental Data: Reexamining Freedman’s Critique. \emph{The Annals of
  Applied Statistics}, 7:295–318.
}
}
