% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_effects.R
\name{BalanceTable}
\alias{BalanceTable}
\title{Covariate Balance Table}
\usage{
BalanceTable(
  data,
  x = NULL,
  treat = "treat",
  xlab = "covariates",
  paired = FALSE,
  percent_vars = c("smd", "pval"),
  format_vars = "stat",
  digits = 1
)
}
\arguments{
\item{data}{A data frame containing covariates for treatment and control
subjects with indicator variable for treatment assignment.}

\item{x}{Character vector containing names of covariates.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{xlab}{Name for table column containing covariates.}

\item{paired}{Whether the data is paired, in which case the SMD and t
statistic will be computed for paired data.}

\item{percent_vars}{Names for variables to format as percentages. Note that
this only changes the printing format, not the actual value of the
variables.}

\item{format_vars}{Names for variables to format as a float. Note that
this only changes the printing format, not the actual value of the
variables.}

\item{digits}{Number of decimals for table statistics.}
}
\value{
A data frame containing the covariate balance table.
}
\description{
Create a table to show covariate balance for a list of specified covariates
\code{x} using the standardized mean difference (SMD) and Welch's t-test
statistic for mean difference and its p-value, with options for paired and
unpaired scenarios. Note that the SMD is Cohen's d for measuring effect size.
}
