% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{BalanceTablePlot}
\alias{BalanceTablePlot}
\title{Covariate Balance Table Plot}
\usage{
BalanceTablePlot(
  data,
  type = c("smd", "stat", "pval"),
  order = c("none", "asc", "desc"),
  smd_threshold = 0.1,
  conf_level = 0.95,
  vline_type = 2,
  scale_log10 = FALSE,
  ylab = "Covariates"
)
}
\arguments{
\item{data}{A data frame containing data for covariate balance table.}

\item{type}{Type of plot.}

\item{order}{Order for the covariates.}

\item{smd_threshold}{Threshold for the SMD.}

\item{conf_level}{Confidence level for inference.}

\item{vline_type}{Line type for vertical lines.}

\item{scale_log10}{Whether to apply a log10 transformation to the x-axis.}

\item{ylab}{Label for y-axis variable.}
}
\value{
A \code{ggplot2::ggplot} horizontal bar plot of the SMD of the
  covariates, or the statistic or p-value for the t-test of the mean
  differences of the covariates.
}
\description{
Creates a horizontal bar plot of the standardized mean difference (SMD) for
for the covariates in a covariate balance table.
}
