% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Binarize}
\alias{Binarize}
\title{Binarize Numeric Vectors}
\usage{
Binarize(x, threshold = 0, strict = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{threshold}{Numeric value for thresholding.}

\item{strict}{Whether inequality condition is strict, i.e., greater than or
greater than or equal to.}
}
\value{
A vector containing the binary indicator variable.
}
\description{
Convert a numeric vector into a binary variable taking values 1 and 0,
depending on whether the vector is greater than a threshold or otherwise.
}
