% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BucketedData}
\alias{BucketedData}
\title{Bucketed Data}
\usage{
BucketedData(
  data,
  vars,
  treat = "treat",
  bucket_var = NULL,
  buckets = NULL,
  seed = NULL
)
}
\arguments{
\item{data}{A data frame containing the treatment variable for the
\code{treat} argument, the numeric variables to be aggregated for the
\code{vars} argument, and optionally, the bucket variable for the
\code{bucket_var} argument.}

\item{vars}{A character vector of the string names of the numeric variables
to be aggregated.}

\item{treat}{String name of treatment variable.}

\item{bucket_var}{String name of bucket variable to bucket by, if it exists.
Note that the bucket variable will always be used for bucketing if
specified, regardless of the value specified for \code{buckets}.}

\item{buckets}{Number of buckets to randomly generate, if no value is
specified for \code{bucket_var}. Number must lie in the range from 1 to
\code{nrow(data)}.}

\item{seed}{Seed for random sampling.}
}
\value{
A data frame containing numeric data aggregated into means by
  buckets, with treatment values and bucket ids as keys.
}
\description{
Data frame of numeric data aggregated into means by buckets.
}
