% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{CausalExpAnalysis}
\alias{CausalExpAnalysis}
\alias{CausalExpAnalysis-package}
\title{CausalExpAnalysis: CausalExpAnalysis R package}
\description{
Functions to perform causal and experiment analysis, i.e., analysis for the causal inference of both experiments and observational studies, within a single framework, with a unifying theme of evaluating covariate balance. Three kinds of samples can be created—unadjusted, weighted, and matched—and two kinds of estimators of treatment effects—an analysis of variance (ANOVA) estimator, i.e., a simple difference-in-means estimator, and an analysis of covariance (ANCOVA) estimator, i.e., a regression adjustment estimator, with either equal or unequal slopes assumption for the covariates. Contains helper functions for statistical inference and modeling, visualization of results, and data manipulation.
}
\author{
\strong{Maintainer}: Shyue-Ming Loh \email{shyuemingloh@gmail.com}

Authors:
\itemize{
  \item Randy Lai \email{randylai@google.com}
  \item Meeyoung Park \email{meeyoung@google.com}
}

}
