% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Center}
\alias{Center}
\title{Center a Variable}
\usage{
Center(x, w = NULL)
}
\arguments{
\item{x}{A numeric vector containing values to center.}

\item{w}{A numeric vector containing weights to construct the centering
mean to subtract from \code{x}. If \code{w} is not \code{NULL}, the
function will stop if the sum of its values is zero, i.e.,
\code{\sum(w) == 0}.}
}
\value{
A centered version of \code{x}.
}
\description{
Center a numeric vector \code{x} by subtracting from its entries a specified
weighted mean:
(1) If the weights \code{w} are not specified, the centering mean is just the
    mean of the vector, i.e., \code{mean(x)}.
(2) If the weights are specified, the centering mean is the following
    weighted mean: \code{sum(w * x) / sum(w)}.
One common application for centering by the above weighted mean is where
\code{w} is the treatment indicator in an analysis of covariance (ANCOVA)
model, so that:
(1) Centering all covariates by the treatment-weighted mean effectively
    centers them by their average in the treatment population, and
(2) Makes the intercept the counterfactual response for the treatment
    population.
}
