% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Combine}
\alias{Combine}
\title{Combine Strings}
\usage{
Combine(x, y, sep = "_", lex = TRUE)
}
\arguments{
\item{x}{The "left" vector of strings.}

\item{y}{The "right" vector of strings.}

\item{sep}{The separator string for concatenation.}

\item{lex}{Whether the combination obeys lexical order or not.}
}
\value{
A vector of all possible concatenations of values from the two
  vectors.
}
\description{
Create all possible combinations of concatenations from two vectors of
strings, with the following variants:
(1) \code{Combine}, the main function for this operation.
(2) \code{Combine0}, the special case of \code{Combine} with \code{sep = ""}.
(3) \code{CombinePaths}, the special case of \code{Combine} with
  \code{sep = "/"}, for combining file paths.
}
