% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_effects.R
\name{CompareModelEffects}
\alias{CompareModelEffects}
\title{Compare Model Effects}
\usage{
CompareModelEffects(
  data,
  y,
  treat,
  post_suffix,
  pre_suffix,
  digits,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame containing an indicator variable for treatment
assignment passed into the \code{treat} argument and post-period and
pre-period metrics for treatment and control subjects specified in
\code{y}.}

\item{y}{Character vector containing names of variables to estimate effects
for. This is a name of the variable without the post-period and pre-period
suffixes passed into, respectively, \code{post_suffix} and {pre_suffix}.
E.g., if the post- and pre-period metrics are, respectively,
\code{"gce_adoption_post"} and \code{"gce_adoption_pre"}, then the name
passed into the \code{y} argument should be \code{"gce_adoption"}.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{post_suffix}{Suffix for the pre-period metric, which will have a
string name equal to \code{paste0(y, pre_suffix)}. Defaults to
\code{"_pre"}.}

\item{verbose}{Whether to print messages.}
}
\value{
A data frame containing the table of statistics and estimated effects
  for the models with Fieller and delta method CIs for the relative lifts of
  the frequentist models.
#' @references
\itemize{
\item Soriano, Jacopo (2019). “Percent Change Estimation in Large Scale
  Online Experiments.” https://arxiv.org/pdf/1711.00562.pdf.
\item Lin, Winston (2013). "Agnostic Notes on Regression Adjustments to
  Experimental Data: Reexamining Freedman’s Critique. \emph{The Annals of
  Applied Statistics}, 7:295–318.
}
}
\description{
Compare the treatment effects for different regression adjustment models,
including the following:
(1) Soriano's (2019) pre-post model
(2) The analysis of covariance (ANCOVA) model with equal slopes and equal
    error variances
(3) Lin's (2013) model, i.e., ANCOVA model with unequal slopes and equal
    error variances
(4) Unequal-variance model, i.e., ANOVA model unequal slopes and unequal
    error variances using a random effects model.
For models (2) through (4), we compute both Fieller and delta method
confidence intervals (CIs) for the relative lifts.
}
