% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{CovariateData}
\alias{CovariateData}
\title{Covariate Data}
\usage{
CovariateData(
  data,
  vars = NULL,
  center = TRUE,
  center_weights = NULL,
  as_matrix = FALSE
)
}
\arguments{
\item{data}{A data frame containing predictors for regression model,
including factors.}

\item{vars}{String names of the covariates in \code{data}. If its value is
\code{NULL}, all the columns in \code{data} will be considered a covariate.}

\item{center}{Whether to center the covariates.}

\item{center_weights}{A numeric vector containing the weights to form the
centering mean. See documentation for \code{Center} for details.}

\item{as_matrix}{Whether to return the covariate data as a matrix. If
\code{FALSE}, a data frame is returned instead.}
}
\value{
A data frame or a matrix containing the covariate values with
  factor contrasts expanded as indicator variables.
}
\description{
Generate a data frame containing the values of the covariates, including
indicator variables for factor levels, using the \code{model.matrix} function
but with the intercept column removed.
}
