% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateFormula}
\alias{CreateFormula}
\title{Create Formula Object from String Names}
\usage{
CreateFormula(y = NULL, x = NULL, collapse = "+")
}
\arguments{
\item{y}{A character vector containing names of the variables on the LHS of
the formula.}

\item{x}{A character vector containing names of the variables on the RHS of
the formula.}

\item{collapse}{A character for separating variables on either side of the
the formula. Defaults to \code{+}.}
}
\value{
A formula object with the LHS and RHS variables specified.
}
\description{
Creates a formula object from string names of variables on left-hand side
(LHS) and right-hand side (RHS) of the formula.
}
