% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DataSlices}
\alias{DataSlices}
\title{Data Slices}
\usage{
DataSlices(data, vars = NULL, sep = ".", global_slice_name = "global")
}
\arguments{
\item{data}{A data frame containing the variables specified in \code{vars}.}

\item{vars}{A character vector containing the string names of the slicing
variables.}

\item{sep}{Separator character between variable name and its unique values,
used as names for the entries of the returned vector. Defaults to the
period.}

\item{global_slice_name}{Name for the global slice.}
}
\value{
A named list of data frames, containing the global slice and other
  slices, sliced by the unique values of the slicing variables in
  \code{vars}.
}
\description{
Given a data frame and a vector of names of slicing variables, return a
named list of data frames containing:
(1) The global slice of the original data frame, i.e., the original data
    frame itself.
(2) Different slices of the data frame, sliced by the unique values of the
    slicing variables.
If no slicing variables are provided, the function creates a list with one
element, i.e., the global slice of the data frame.
}
