% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{DeltaRatioCI}
\alias{DeltaRatioCI}
\title{Delta Method Confidence Interval for the Ratio of Means}
\usage{
DeltaRatioCI(...)
}
\arguments{
\item{...}{Parameters passed into the \code{RatioCI} function. For details,
see parameter description for \code{RatioCI} function.}
}
\value{
A data frame containing the lower and upper bounds of the delta
}
\description{
Wrapper of the \code{RatioCI} function to compute the delta method
confidence interval (CI) for a ratio of two means, given two sample mean
estimates and their covariance matrix. See documentation for the
the \code{RatioCI} function.
}
