% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DistinctValueFreqTable}
\alias{DistinctValueFreqTable}
\title{Distinct Value Frequency Table}
\usage{
DistinctValueFreqTable(data, x, unit)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Variable that we would like to obtain frequencies for its distinct
values.}

\item{unit}{Variable specifying the IDs for the frequency units.}
}
\value{
A data frame containing frequencies for each number of distinct
  values of the variable \code{x}.
}
\description{
Create a table of frequencies of units having k number of distinct values of
a specific variable \code{x}.
}
