% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{EffectiveSampleSize}
\alias{EffectiveSampleSize}
\title{Effective Sample Size}
\usage{
EffectiveSampleSize(w, all_stats = FALSE)
}
\arguments{
\item{w}{A numeric vector of weights for the weighted mean.}

\item{all_stats}{Whether to return all the relevant statistics. If
\code{TRUE}, a data frame will be returned with the effective sample size,
the nominal sample size, and the relative efficiency. Otherwise, the scalar
value of the effective sample size will be returned.}
}
\value{
Either a data frame containing the effective and nominal sample
  sizes with their ratio, or a scalar value of the effective sample size,
  depending on the \code{all_stats} argument.
}
\description{
Compute the effective sample size based on the relative efficiency of the
estimation of a weighted mean.
}
