% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{EffectSize}
\alias{EffectSize}
\title{Effect Size from Mean Difference}
\usage{
EffectSize(
  m1,
  m2 = 0,
  s1,
  s2 = NULL,
  n1 = NULL,
  n2 = NULL,
  method = c("cohens", "sdiff")
)
}
\arguments{
\item{m1}{Mean of first sample.}

\item{m2}{Mean of second sample.}

\item{s1}{Standard deviation of first sample.}

\item{s2}{Standard deviation of second sample.}

\item{n1}{Size of first sample.}

\item{n2}{Size of second sample.}

\item{method}{One of two choices, whether \code{cohens} for Cohen's d or
\code{sdiff} for the standardized difference of}
}
\value{
Effect size estimate from mean difference.
\itemize{
\item Cohen, Jacob (1988). \emph{Statistical Power Analysis for the
   Behavioral Sciences,} 2nd ed. New Jersey: Lawrence Erlbaum Associates.
\item Rosenbaum, Paul R. and Donald B. Rubin DB (1985). "Constructing a
   Control Group Using Multivariate Matched Sampling Methods That
   Incorporate the Propensity Score." \emph{The American Statistician.}
   39: 33-38.
}
}
\description{
Compute the effect size as a measure of practical significance by
standardizing a mean difference. Implements two conventions:
(1) Cohen's d, which divides the mean difference by the pooled standard
    deviation of the two samples, as shown in, e.g., Cohen (1988).
(2) Standardized Difference, which divides the mean difference by the
    square root of the average of two sample variances, as proposed by
    Rosenbaum and Rubin (1985).
}
