% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{EffectSizeCI}
\alias{EffectSizeCI}
\title{Effect Size Confidence Interval}
\usage{
EffectSizeCI(d, n1, n2 = NULL, conf_level = 0.95, ci_vars = c("lb", "ub"))
}
\arguments{
\item{d}{A numeric vector containing values of the Hedges' d standardized
mean difference (SMD) effect size, or Cohen's d SMD effect size for large
sample sizes.}

\item{n1}{Size of the first sample, which is required.}

\item{n2}{Size for the second sample, which is optional. If \code{NULL}, then
it is set equal to the size \code{n1} of the first sample.}

\item{conf_level}{Confidence level for interval.}

\item{ci_vars}{Names for the lower and upper bounds of the CI in the output
data frame.}
}
\value{
A data frame containing the asymptotic CIs for Hedges' d and Cohen's
  d SMD effect sizes.
}
\description{
Computes the confidence interval (CI) for the Cohen's d standardized mean
difference (SMD) effect size, i.e., the estimated effect divided by the
estimated pooled standard deviation, based on its asymptotic normal
distribution, as shown on p. 86 of Hedges and Olkin (1985). To be accurate,
the CI formula derived by Hedges and Olkin (1985) is for Hedges' d, which
multiplies Cohen's d by a bias correction factor. However, the bias
correction factor is close to 1 for large sample sizes, so that both Hedges'
d and Cohen's d and their CIs are asymptotically equivalent. Hence, the CI
can be used for Cohen's d for large sample sizes.
}
\references{
\itemize{
\item Hedges, Larry and Ingram Olkin (1985). \emph{The Statistical Methods
  for Meta-analysis}. Orlando: Academic Press.
}
}
