% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{EffectSizeSD}
\alias{EffectSizeSD}
\title{Standard Deviation for Computing Effect Size of Mean Difference}
\usage{
EffectSizeSD(s1, s2, n1 = NULL, n2 = NULL, method = c("pooled", "average"))
}
\arguments{
\item{s1}{Standard deviation of first sample.}

\item{s2}{Standard deviation of second sample.}

\item{n1}{Size of first sample.}

\item{n2}{Size of second sample.}

\item{method}{One of two choices, whether \code{pooled} for the pooled
standard deviation or \code{average} for the square root of the average
sample variance.}
}
\value{
Standard deviation for effect size calculation.
}
\description{
Computes the standard deviation to divide a mean difference for obtaining
an effect size, using either the pooled standard deviation, which gives us
the Cohen's d effect size, or the average standard deviation, which gives us
the standardized difference effect size. See documentation for
\code{EffectSize} for details.
}
