% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_effects.R
\name{EffectTable}
\alias{EffectTable}
\title{Treatment Effect Table}
\usage{
EffectTable(
  data,
  metrics,
  paired = FALSE,
  var_equal = FALSE,
  treat = "treat",
  match = "match",
  type = c("both", "abs", "rel"),
  reff_ci_method = c("fieller", "delta", "delta_log"),
  conf_level = 0.95,
  sig_test = TRUE,
  eff_size = TRUE,
  digits = 1
)
}
\arguments{
\item{data}{A data frame containing metrics for treatment and control
subjects with indicator variable for treatment assignment. Binary metrics
can be included as well.}

\item{metrics}{Character vector containing names of metrics.}

\item{paired}{Whether the data is paired, in which case the SMD and t
statistic will be computed for paired data.}

\item{var_equal}{Whether to assume unequal variances for the metrics across
treatment and control groups. Similar to the \code{var.equal} argument in
the \code{stats::t.test} function. Applies only when \code{paired = FALSE}.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{match}{Name for variable for matching paired subjects.}

\item{type}{The type of effect(s) to show in the table, whether an absolute
effect, a relative effect, or both.}

\item{reff_ci_method}{A string for the relative effect CI method, either
\code{"fieller"} for the Fieller CI, \code{"delta"} for the delta method
without log transformation of the ratio, or \code{"delta_log"} for the
delta method with log transformation of the ratio. See documentation for
the \code{RatioCI} function for details.}

\item{conf_level}{Confidence level for inference.}

\item{sig_test}{Whether to show the inference statitics for a significance
test, i.e., the standard error of the effect estimate, the t statistic, and
the p-value.}

\item{eff_size}{Whether to show the statitics for measuring effect size,
namely, the joint standard deviation for the two samples and the
standardized mean difference (SMD), i.e., the mean difference divided by
the joint standard deviation.}

\item{digits}{Number of decimals for table statistics.}

\item{match_counts}{Whether to show the matched counts \code{x00},\code{x01},
\code{x10}, and \code{x11} in the table.}
}
\value{
A data frame containing the effect table for binomial proportions.
}
\description{
Create a table to show mean differences between the treatment and control
groups on any set of variables, which we will refer to as metrics. The
metrics can be experiment outcomes or pre-treatment covariates. In the former
case, the mean difference is the average treatment effect for the outcome;
in the latter case, the mean difference measures the covariate imbalance
between the two arms. User has the option to measure two types of effects:
(1) Absolute effect, where the t test is used to construct CIs and optionally
    show the inference statistics for the test, including t statistic and
    p-value, and the standardized mean difference with corresponding
    observation standard deviation, as a measure of effect size.
(2) Relative effect, where the CIs are constructed using either the Fieller
    or delta method.
The metrics passed into the function can include categorical variables,
either factors or character vectors, in which case, the variable is expanded
into a set of dummy variables using the default contrasts available in R.
There is also an option to adjust the SEs for paired data.
}
