% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_effects.R
\name{EffectTableProp}
\alias{EffectTableProp}
\title{Treatment Effect Table for Binomial Proportions}
\usage{
EffectTableProp(
  data,
  metrics,
  paired = FALSE,
  weight = "n",
  treat = "treat",
  match = "match",
  type = c("abs", "rel", "both"),
  conf_level = 0.95,
  sig_test = FALSE,
  digits = 1,
  match_counts = FALSE
)
}
\arguments{
\item{data}{A data frame containing binary metrics for treatment and control
subjects with indicator variable for treatment assignment.}

\item{metrics}{Character vector containing names of metrics.}

\item{paired}{Whether the data is paired, in which case the SMD and t
statistic will be computed for paired data.}

\item{weight}{Name of variable containing subject weights.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{match}{Name for variable for matching paired subjects.}

\item{type}{The type of effect(s) to show in the table, whether an absolute
effect, a relative effect, or both.}

\item{conf_level}{Confidence level for inference.}

\item{sig_test}{Whether to show the inference statitics for a significance
test, i.e., the standard error of the effect estimate, the t statistic, and
the p-value.}

\item{match_counts}{Whether to show the matched counts \code{x00},\code{x01},
\code{x10}, and \code{x11} in the table.}
}
\value{
A data frame containing the effect table for binomial proportions.
}
\description{
Create a table to show the average treatment effect for binomial proportions
of binary metrics, showing the estimated mean difference with standard error,
z statistic, and p-value, with options for paired and unpaired proportions
}
