% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{EstimateTablePlot}
\alias{EstimateTablePlot}
\title{Mean Estimate and Treatment Effect Table Plots}
\usage{
EstimateTablePlot(
  data,
  type = c("eff", "reff", "stat", "pval", "smd", "est"),
  plot_type = c("bar", "point"),
  order = c("none", "asc", "desc"),
  conf_level = 0.95,
  ref_slice = "all",
  vline_type = 2,
  text = FALSE,
  scale_log10 = FALSE,
  scale_percent = FALSE,
  y = "metric",
  ylab = "Metrics",
  smd_ci = FALSE,
  smd_threshold = 0.02,
  sig_level = 0.05,
  power = 0.8,
  n_vars = c("nT", "nC"),
  ...
)
}
\arguments{
\item{data}{A data frame containing data for covariate balance table.}

\item{type}{The \code{ggplot2} geom to use, whether \code{"bar"} for a bar
plot, or \code{"point"} for a point plot.}

\item{order}{Order for the covariates.}

\item{conf_level}{Confidence level for inference.}

\item{ref_slice}{The value for the \code{slice} variable in the data for an
\code{est} plot, for use in determining the x-axis value of the vertical
line.}

\item{vline_type}{Line type for vertical lines.}

\item{text}{Whether to show values as text labels.}

\item{scale_log10}{Whether to apply a log10 transformation to the x-axis.}

\item{y}{Name of y variable. Defaults to \code{"metric"}.}

\item{ylab}{Label for y-axis variable.}

\item{smd_threshold}{Threshold for minimum effect size expressed by the
standardized mean difference (SMD). If the string \code{"mdes"} is passed
in instead, the minimum detectable effect size (MDES) will be used as the
SMD threshold.}

\item{sig_level}{Significance level for computing the MDES.}

\item{power}{Power for computing the MDES.}

\item{n_vars}{A character vector of length 2 containing the names of the
variables in \code{data} containing, respectively, the sample sizes for
the treatment and control groups.}

\item{...}{Additional arguments for the \code{HBarPlot} function.}
}
\value{
A \code{ggplot2::ggplot} horizontal bar plot of statistic or p-value
  for the t-test of the mean differences of the covariates.
}
\description{
Creates a horizontal bar or point plot of one of the following measures:
(1) A mean estimate of a response, specified using \code{type = "est"}.
(2) The absolute effect, specified using \code{type = "eff"}.
(3) The relative ffect, specified using  \code{type = "reff"}.
(4) The z-statistic of the absolute effect, specified using
    \code{type = "stat"}.
(5) The p-value of the absolute effect, specified using \code{type = "pval"}.
(6) The standardized mean difference (SMD) as a measure of effect size,
    specified using \code{type = "smd"}.
}
\details{
Note that this function was originally named \code{EffectTablePlot} but has
been renamed to \code{EstimateTablePlot}, because it has been generalized
to include non-effect measures, e.g., the mean estimate of a response.
However, for backward compatibility's sake, the \code{EffectTablePlot}
function name works for the same function as well.
}
