% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_subgroups.R
\name{EstimationStats}
\alias{EstimationStats}
\title{Estimation Statistics}
\usage{
EstimationStats(
  data,
  y,
  treat = NULL,
  est = NULL,
  se = NULL,
  conf_level = 0.95,
  resp_name = "est",
  eff_name = "eff"
)
}
\arguments{
\item{data}{A data frame containing the response metric and the slicing
covariates, and optionally, the treatment variable, if the ATE is
estimated.}

\item{y}{A character vector of names of metric.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{est}{The value of an external model estimate.}

\item{se}{The value of the standard error (SE) of the external model
estimate.}

\item{conf_level}{Confidence level for inference.}

\item{resp_name}{The name of the estimated mean response in the output table.}

\item{eff_name}{The name of the estimated ATE in the output table.}
}
\value{
A single-row data frame containing the estimation statistics.
}
\description{
A helper for the \code{SubgroupAnalysis} function to compute the estimation
statistics for each slice. Returns the table as a single-row data frame, with
the following statistics, depending on whether the mean response or the
average treatment effect (ATE) is estimated:
(1) Mean Response: Observation count \code{n}, the sample mean \code{mean},
    the estimated mean \code{est}, the confidence interval (CI) lower and
    upper bounds \code{lb} and \code{ub}, and the standard error (SE) of the
    estimate.
(2) Average Treatment Effect: Observation counts for Control and Treatment
    groups \code{nC} and \code{nT}, the sample means for the Control and
    Treatment groups \code{mC} and \code{mT}, the sample mean difference
    between the Treatment and Control groups \code{diff}, the estimated
    average treatment effect \code{eff}, its confidence interval (CI) lower
    and upper bounds \code{lb} and \code{ub}, and the standard error (SE) of
    the treatment effect estimate.
}
