% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ExtractLinearModelEffect}
\alias{ExtractLinearModelEffect}
\title{Extract Treatment Effect Estimate from Linear Model}
\usage{
ExtractLinearModelEffect(
  m,
  treat = "treat",
  hc = c("none", "hc0", "hc1", "hc2", "hc3", "hc4"),
  conf_level = 0.95,
  rel_eff = TRUE,
  reff_ci_method = c("fieller", "delta", "delta_log"),
  design = NULL
)
}
\arguments{
\item{m}{A \code{stats::lm}, \code{lme4::lmer}, \code{lmerTest::lmer}, or
\code{survey::svyglm} model object.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{hc}{Option for computing heteroskedasticity-corrected SEs using
\code{car::hccm}. Defaults to \code{"none"}. The other values are
\code{"hc0"}, \code{"hc1"}, \code{"hc2"}, \code{"hc3"}, and \code{"hc4"}.
See documentation for \code{car::hccm} for details. Applies only to pure
\code{stats::lm}. Doesn't apply to \code{survey::svyglm},
\code{lme4::lmer}, or \code{lmerTest::lmer} objects.}

\item{conf_level}{Confidence level for inference.}

\item{rel_eff}{Whether to compute the relative effect and its CI.}

\item{reff_ci_method}{A string for the relative effect CI method, either
\code{"fieller"} for the Fieller CI, \code{"delta"} for the delta method
without log transformation of the ratio, or \code{"delta_log"} for the
delta method with log transformation of the ratio. See documentation for
the \code{RatioCI} function for details.}

\item{design}{An object of class \code{svydesign}, optionally included to
pass into the \code{survey::svyvar} function to compute the survey
population variance for the \code{stats::lm} model fits, if this argument
is not NULL.}
}
\value{
A single-row data frame containing the regression-adjusted absolute
  effect with corresponding inference statistics, and optionally, the
  relative effect and with its CI.
}
\description{
Extract the estimated absolute treatment effect from a linear model with a
treatment indicator variable, fitted using either
\code{stats::lm}, \code{lme4::lmer}, \code{lmerTest::lmer}, or
\code{survey::svyglm}. Compute corresponding statistics, including:
(1) Baseline and treatment response values, assuming that the linear model
    intercept gives the correct baseline value due to the appropriate
    centering of the covariates.
(2) The standard error and t statistic for the effect,
(3) The p-value for the significance test of the effect,
(4) The confidence interval (CI) for the effect, and
(5) The population standard deviation and standardized mean difference.
Optionally compute the relative effect and its CI using the Fieller or
delta method.
}
