% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ExtractModelStats}
\alias{ExtractModelStats}
\title{Extract Model Stats}
\usage{
ExtractModelStats(
  data,
  formula,
  model = c("ancova", "did"),
  method,
  unit = NULL,
  weights = NULL,
  design = NULL,
  single_obs,
  treat = "treat",
  hc = c("none", "hc0", "hc1", "hc2", "hc3", "hc4"),
  conf_level = 0.95,
  rel_eff = TRUE,
  reff_ci_method = c("fieller", "delta", "delta_log"),
  output_models = FALSE,
  trim_fit = TRUE
)
}
\arguments{
\item{data}{A data frame containing metrics for treatment and control
subjects, an indicator variable for treatment assignment and optionally
additional covariates and an integer variable for matching of pairs.}

\item{formula}{A formula object to pass into \code{stats::lm},
\code{lmerTest::lmer}, or \code{survey::svyglm}.}

\item{model}{The type of regression model, whether \code{ancova} for a
unconstrained ANCOVA model or \code{did} for a difference-in-differences
model.}

\item{method}{Whether \code{"lm"} for \code{stats::lm} or \code{"lmer"} for
\code{lmerTest::lmer}.}

\item{weights}{Name of variable containing weights for use in fitting the
the linear model.}

\item{design}{An object of class \code{svydesign}, optionally included to
pass into the \code{survey::svyglm} function to fit the linear model, so
we do not need to create a new \code{svydesign} object in the function
call.}

\item{single_obs}{Whether to allow for a single observation within a
random-effect level when \code{lmerTest::lmer} is run.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{hc}{Option for computing heteroskedasticity-corrected SEs using
\code{car::hccm}. Defaults to \code{"none"}. The other values are
\code{"hc0"}, \code{"hc1"}, \code{"hc2"}, \code{"hc3"}, and \code{"hc4"}.
See documentation for \code{car::hccm} for details. Applies only to pure
\code{stats::lm}. Doesn't apply to \code{survey::svyglm},
\code{lme4::lmer}, or \code{lmerTest::lmer} objects.}

\item{conf_level}{Confidence level for inference.}

\item{rel_eff}{Whether to compute the relative effect and its CI.}

\item{reff_ci_method}{A string for the relative effect CI method, either
\code{"fieller"} for the Fieller CI, \code{"delta"} for the delta method
without log transformation of the ratio, or \code{"delta_log"} for the
delta method with log transformation of the ratio. See documentation for
the \code{RatioCI} function for details.}

\item{output_models}{Whether to return the regression model objects together
in addition to the table of model stats. In this case, a list will be
returned containing the model stats in the first entry and a list of model
objects in the second entry.}

\item{trim_fit}{Whether to apply the \code{TrimLinearModelFit} function to
a \code{stats::lm} or \code{survey::svyglm} object.}
}
\value{
A data frame of model stats if \code{output_models = FALSE} and a
  list containing a data frame of model stats and a list of model objects
  if \code{output_models = TRUE}.
}
\description{
Function used \code{AdjustedEffectTable} to extract model stats using the
a combination of the \code{FitLinearModel} and
\code{ExtractLinearModelEffect} functions.
}
