% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{FiellerCI}
\alias{FiellerCI}
\title{Fieller Confidence Interval for the Ratio of Means}
\usage{
FiellerCI(...)
}
\arguments{
\item{...}{Parameters passed into the \code{RatioCI} function.}
}
\value{
A data frame containing the lower and upper bounds of the
  Fieller CI.
}
\description{
Wrapper of the \code{RatioCI} function to compute the Fieller confidence
interval (CI) for a ratio of two means, given two sample mean estimates and
their covariance matrix. Implemented in a "tidy" fashion by returning the
results as a data frame. The arguments \code{a, b, v11, v22, v12, s} can be
numeric vectors with multiple entries.
}
