% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{FiellerCIMatchedProp}
\alias{FiellerCIMatchedProp}
\title{Fieller Confidence Interval for the Ratio of Matched Proportions}
\usage{
FiellerCIMatchedProp(a, b, c, d, conf_level = 0.95, ...)
}
\arguments{
\item{a}{A numeric vector of pair counts where both variables are 0}

\item{b}{A numeric vector of pair counts where first variable is 0 and the
second 1.}

\item{c}{A numeric vector of pair counts where the first variable is 1 and
the second 0.}

\item{d}{A numeric vector of pair counts where both variables are 1.}

\item{conf_level}{Confidence level for interval.}

\item{...}{Additional arguments to pass into \code{FiellerCI}}
}
\value{
A data frame containing the lower and upper bounds of the
  Fieller confidence interval.
}
\description{
Wrapper of the \code{FiellerCI} function to compute the Fieller confidence
interval for a ratio of two matched or paired proportions, given the counts
of the 2x2 contingency table for their paired values. The arguments
\code{a, b, c, d} can be numeric vectors with multiple entries. Covariance
between two matched proportions are computed from the relation:
Cov(p1, p2) = (Var(p1) + Var(p2) - Var(p1 - p2)) / 2
}
