% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{FiellerCIProp}
\alias{FiellerCIProp}
\title{Fieller Confidence Interval for the Ratio of Proportions}
\usage{
FiellerCIProp(p1, p2, n1, n2, r = 0, conf_level = 0.95, ...)
}
\arguments{
\item{p1}{Estimate for the numerator of the ratio.}

\item{p2}{Estimate for the denominator of the ratio.}

\item{n1}{Size for the first sample.}

\item{n2}{Size for the second sample.}

\item{r}{Estimate of correlation between p1 and p2.}

\item{conf_level}{Confidence level for interval.}

\item{...}{Additional arguments to pass into \code{FiellerCI}}
}
\value{
A data frame containing the lower and upper bounds of the
  Fieller confidence interval.
}
\description{
Wrapper of the \code{FiellerCI} function to computes the Fieller confidence
interval for a ratio of two proportions, given two sample proportions, their
sample sizes, and their mutual correlation. The arguments
\code{p1, p2, n1, n2, r} can be numeric vectors with multiple entries.
}
