% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{FitCausalModels}
\alias{FitCausalModels}
\title{Fit Causal Models}
\usage{
FitCausalModels(
  data,
  y,
  treat,
  vars,
  center = FALSE,
  method = c("causal_forest", "bartc", "tmle", "rlasso"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame containing variables for the model.}

\item{y}{A vector of names of response variables.}

\item{treat}{The name of the treatment variable.}

\item{vars}{A vector of names of covariates.}

\item{center}{Whether to center the covariates at the sample mean.}

\item{method}{Method for the model, either \code{"causal_forest"},
\code{"bartc"}, \code{"tmle"}, or \code{"rlasso"} for, respectively, the
four models listed above.}

\item{verbose}{Whether to print messages and elapsed compute times.}
}
\value{
A list of model objects returned by one of the four models described
  above.
\itemize{
\item Athey, Susan, Julie Tibshirani, and Stefan Wager (2019). "Generalized
  Random Forests." \emph{Annals of Statistics}, 47(2): 1148-1178.
\item van der Laan, Mark J. and Daniel Robun (2006). "Targeted Maximum
  Likelihood Learning." \emph{The International Journal of Biostatistics},
  2(1).
\item Chernozhukov, Victor, Chris Hansen, and Martin Spindler (2016). "hdm:
  High-dimensional Metrics." \emph{The R Journal}, 8(2): 185-199.
}
}
\description{
A wrappper function to fit multiple causal models to estimate average
treatment effects, using one of the following functions:
(1) \code{grf::causal_forest} to fit a causal forest using the generalized
    random forests framework of Athey et al. (2019).
(2) \code{bartCause::bartc} to fit a causal Bayesian Additive Regression
    Tree (BART).
(3) \code{tmle::tmle} to fit a causal model using targeted minimum loss-based
    estimation (TMLE) method of van der Laan and Rubin (2006).
(4) \code{hdm::rlassoATE} to fit a "rigorous" lasso model described in
    Chernozhukov et al. (2016).
}
