% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{FitLinearModel}
\alias{FitLinearModel}
\title{Fit Linear Model}
\usage{
FitLinearModel(
  data,
  formula,
  random = NULL,
  correlation = NULL,
  method = c("lm", "svyglm", "lmer", "lme", "gls"),
  weights = NULL,
  design = NULL,
  single_obs = FALSE,
  trim_fit = FALSE
)
}
\arguments{
\item{data}{A data frame containing variables for the linear model.}

\item{formula}{A formula object to pass into \code{stats::lm},
\code{lmerTest::lmer}, or \code{survey::svyglm}.}

\item{random}{A formula for the random effect, or a list of random effect
formulas, used only when the \code{nlme::lme} function is used.}

\item{method}{Whether \code{"lm"} for \code{stats::lm}, \code{"svyglm"} for
\code{survey::svyglm}, \code{"lmer"} for \code{lmerTest::lmer},
\code{"lme"} for \code{nlme::lme}, and \code{"gls"} for \code{nlme::gls}.}

\item{weights}{Name of variable containing weights for use in fitting the
the linear model.}

\item{design}{An object of class \code{svydesign}, optionally included to
pass into the \code{survey::svyglm} function to fit the linear model, so
we do not need to create a new \code{svydesign} object in the function
call.}

\item{single_obs}{Whether to allow for a single observation within a
random-effect level when \code{lmerTest::lmer} is run.}

\item{trim_fit}{Whether to apply the \code{TrimLinearModelFit} function to
a \code{stats::lm} or \code{survey::svyglm} object.}
}
\value{
A \code{stats::lm} or  \code{lmerTest::lmer} model object.
}
\description{
Fit a linear fixed-effects model using \code{stats::lm}, a linear
mixed-effects model using \code{lme4::lmer}, or a survey-weighted linear
model using \code{survey::svyglm}.
}
