% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_samples.R
\name{FitPropensityScoreModel}
\alias{FitPropensityScoreModel}
\title{Fit Propensity Score Model}
\usage{
FitPropensityScoreModel(
  data,
  treat = "treat",
  covariates = NULL,
  method = c("ps", "gbm_es", "gbm_ks", "xgboost_es", "xgboost_ks", "cbps", "cbps_np",
    "cal", "ebal"),
  ps_model = c("glm", "gam", "glmnet"),
  estimand = c("att", "ate", "atu", "atc", "ato"),
  score = "score",
  weights = "weights",
  alpha = 0,
  lambda = c("min", "1se"),
  normalized = FALSE
)
}
\arguments{
\item{data}{A data frame containing the treatment variable and covariates
for fitting a model for propensity score or balancing weights.}

\item{treat}{The name of the treatment variable.}

\item{covariates}{Character vector containing names of covariates.}

\item{method}{The method for fitting the propensity score model, as listed
earlier.}

\item{estimand}{The causal estimand for fitting the propensity score model,
if applicable, or for fitting the balancing weight model. See values
listed earlier in the above documentation.}

\item{score}{Name of the variable in the returned data that contains the
estimated propensity score.}

\item{weights}{Name of the variable in the returned data that contains the
estimated propensity score weights or balancing weights.}

\item{alpha}{The elastic-net mixing parameter between 0 and 1, with
\code{alpha = 0} for ridge regression and \code{alpha = 1} for lasso,
applicable for \code{method = "ps"} and \code{ps_model = "glmnet"}.}

\item{lambda}{The criterion to select the lambda of the elastic net, either
\code{"min"} for the minimum lambda value or \code{"1se"} for lambda plus
one standard error, applicable for \code{method = "ps"} and
\code{ps_model = "glmnet"}.}
}
\value{
An object of class \code{psm}, described earlier.
\itemize{
\item Chan, Kwun Chuen Gary, Sheung Chi Phillip Yam, and Zheng Zhang (2016).
  "Globally Efficient Non-parametric Inference of Average Treatment Effects
  by Empirical Balancing Calibration Weighting." \emph{Journal of the Royal
  Statistical Society: Series B (Statistical Methodology)}, 78(3): 673-700.
\item Hainmueller, Jens (2012). "Entropy Balancing for Causal Effects: A
  Multivariate Reweighting Method to Produce Balanced Samples in
  Observational Studies." \emph{Political Analysis}, 20(1): 25-46.
\item Imai, Kosuke and Marc Ratkovic (2014). "Covariate Balancing Propensity
  Score." \emph{Journal of the Royal Statistical Society B}, 76(1): 243-263.
\item Li, Fan, Kari Lock Morgan, and Alan M. Zaslavsky (2018). "Balancing
  Covariates via Propensity Score Weighting." \emph{Journal of the American
  Statistical Association}, 113(521): 390-400.
}
}
\description{
Fit a propensity score model given treatment variable and covariates. Note
that we use the phrase "propensity score model" in a very broad sense
to include even models where the propensity score are not estimated directly
but can be derived from covariate-balancing weights under suitable
conditions. There are three broad categories for these models:
(1) Models that estimate the propensity score directly using a criterion
    that is agnostic to a specific causal estimand, e.g., maximizing a
    binomial likelihood or a penalized version of the likelihood.
(2) Models that estimate the propensity score directly using a criterion
    specific to a causal estimand, e.g., a covariate-balancing criterion
    optimizing for the average treatment effect on the entire population
    (ATE) or the average treatment effect on the treated population (ATT).
(3) Models that estimate the covariate-balancing weights directly without
    estimating a propensity score. These weights are optimized specifically
    for a causal estimand, e.g., the ATE or the ATT. They can be used to
    derive a propensity score if the causal estimand is the ATE. But if the
    the causal estimand is the ATT, we can derive the propensity score only
    for the untreated/control population. Similarly, if the causal estimand
    is the ATU/ATC, we can derive the propensity score only for the treated
    population.
  We use the \code{method} and \code{ps_model} arguments to categorize the
above three broad categories of models. The following \code{method} covers
models that estimate the propensity score directly using a criterion
not specific to a causal estimand:
(1) \code{ps}, covers models that estimate the propensity score directly
    using a criterion not specific to a causal estimand: The different models
    used under this method are identified using the \code{ps_model} argument.
  The following values for the \code{method} argument cover models that
estimate a propensity score using a criterion specific to a causal estimand:
(2) \code{gbm_es}, which fits a gradient-boosted tree model using
    \code{twang::ps}, with \code{version = "gbm"} and the effect size (ES)
    stopping method using \code{stop.method = "es.mean"}.
(3) \code{gbm_ks}, which fits a gradient-boosted tree model using
    \code{twang::ps}, with \code{version = "gbm"} and the Kolmogorov-Smirnov
    (KS) stopping method using \code{stop.method = "ks.max"}.
(4) \code{xgboost_es}, which fits an XGBoost model using \code{twang::ps},
    with \code{version = "xgboost"} and the ES stopping method using
    \code{stop.method = "es.mean"}.
(5) \code{xgboost_ks}, which fits an XGBoost model using \code{twang::ps},
    with \code{version = "xgboost"} and the KS stopping method using
    \code{stop.method = "ks.max"}.
(6) \code{cbps}, the covariate-balancing propensity score (CBPS) method of
    Imai et al. (2014), which maximizes covariate balance instead of the
    binomial likelihood of the propensity score model.
(7) \code{cbps_np}, the non-parametric version of the CBPS for continuous
    treatment but applied to binary treatments.
  The following values for the \code{method} argument cover models that
estimate the covariate-balancing weights directly without estimating a
propensity score:
(8) \code{cal}, the calibration weights of Chan et al. (2017).
(9) \code{ebal}, the entropy-balancing weights of Hainmuller (2012).
  For \code{method = "ps"}, the different propensity-score models are
identified using the \code{ps_model} argument, which has the following
values:
(1) \code{glm}, which fits a logistic regression model using
    \code{stats::glm} with \code{family = binomial}.
(2) \code{gam}, which fits a logistic regression model with penalized
    iterated reweighted least squares (PIRLS) algorithm using
    \code{mgcv::gam} with \code{family = binomial}.
(3) \code{glmnet}, which fits an elastic-net logistic regression model using
    \code{glmnet::cv.glmnet} with \code{family = "binomial"} and the
    elastic-net mixing parameter \code{alpha} specified as an argument.
  The propensity score weights are estimated for the following causal
estimands, passed into the \code{estimand} argument:
(1) \code{att}, the average treatment effect on the treated (ATT), available
    for all methods, except for \code{cbps_np}.
(2) \code{ate}, the average treatment effect on the entire population (ATE),
    available for all methods except \code{ebal}.
(3) \code{atu} or \code{atc}, the average treatment effect on the untreated
    or control (ATU or ATC), available for all methods, except for
    \code{cbps_np}.
(4) \code{ato}, the average treatment effect on the overlap population (ATO),
    available only for the \code{ps} method, as discussed in Li et al.
    (2018).
  The function returns the a \code{psm} object containing the following:
(1) \code{data}, the original data with the estimated propensity score and
    propensity score weights added to it.
(2) \code{model}, the model fit object.
(3) \code{method}, the method used to estimate the propensity score or
    propensity score weights.
(4) \code{ps_model}, the propensity score model used for the \code{ps}
    method.
(4) \code{estimand}, the causal estimand for estimating the score, if
    applicable, or for deriving the propensity score weights.
(5) \code{treat}, the name of the treatment variable.
(6) \code{score}, the name of the propensity score variable.
(7) \code{weights}, the name of the propensity score weight or
    balancing weight variable.
}
