% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FormatTable}
\alias{FormatTable}
\title{Apply Numeric and Percent Printing Formats to Columns in a Data Frame}
\usage{
FormatTable(data, format_vars = NULL, percent_vars = NULL, digits = 1)
}
\arguments{
\item{data}{A data frame.}

\item{format_vars}{A character vector containing names of columns to apply
numeric printing format to.}

\item{percent_vars}{A character vector containing names of columns to apply
percent printing format to.}

\item{digits}{The number of digits for the printing numbers.}
}
\value{
The data frame passed into the function with the
  numeric and percent printing formats applied to the selected columns.
}
\description{
Apply numeric printing format from \code{formattable::formattable} and
percent printing format from \code{formattable::percent} to a selection
of columns in a data frame.
}
