% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{FreqTable}
\alias{FreqTable}
\title{Frequency Table}
\usage{
FreqTable(
  data,
  group,
  unit = NULL,
  unit_label = NULL,
  over = NULL,
  prop = TRUE,
  ci = FALSE,
  conf_level = 0.95,
  digits = 1,
  format = "f"
)
}
\arguments{
\item{data}{A data frame.}

\item{group}{A character vector containing names of variables to group by to
obtain frequencies.}

\item{unit}{Name of variable containing the key to the observation units.}

\item{unit_label}{The label for the observation units in the frequency table.}

\item{over}{A character vector containing names of variables to group by to
obtain totals to divide frequencies by; could potentially be different
what is specified in \code{group}.}

\item{prop}{Whether to include the frequency proportions.}

\item{ci}{Whether to include the confidence interval for the proportions.}

\item{conf_level}{Confidence level for interval.}

\item{digits}{Number of digits to print for percent variables.}

\item{format}{The \code{format} argument passed into the
\code{formattable::percent} function.}
}
\value{
A data frame containing the frequency table.
}
\description{
Create a table of frequency counts with, optionally, frequency proportions
and proportion confidence intervals.
}
