% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{GroupTimeSeriesPlots}
\alias{GroupTimeSeriesPlots}
\title{Group Time Series Plots}
\usage{
GroupTimeSeriesPlots(
  data,
  y,
  x,
  group = "group",
  type = c("line", "smooth"),
  FUN = mean,
  facet = NULL,
  facet_all = NULL,
  facet_nrow = 1,
  facet_scales = "fixed",
  scale_percent = FALSE
)
}
\arguments{
\item{data}{A data frame containing data for the time series plots.}

\item{y}{A character vector of names for the time series metrics.}

\item{x}{Name of the variable in the x-axis, typically a date variable.}

\item{group}{Name of the group variable.}

\item{type}{Whether the plot is a line plot or a smooth plot.}

\item{FUN}{The function for aggregating \code{y} variables by group for line
plots only. Defaults to \code{mean}. Note that for smooth plots, the only
aggregation function available is the mean.}

\item{facet}{Name of facet variable.}

\item{facet_all}{Facet value for the entire dataset, if this argument is
not null. The aggregated data will be appended with a dataset aggregated
over all observations in the data with the facet variable value equal to
\code{facet_all}.}

\item{facet_nrow}{The value passed into the \code{nrow} argument of the
\code{ggplot2::facet_wrap} function.}

\item{facet_scales}{The value passed into the \code{scales} argument of the
\code{ggplot2::facet_wrap} function.}
}
\value{
A list of \code{ggplot2::ggplot} objects containing the line or
  smooth plots.
}
\description{
Creates a list of group-aggregated time series plots for specified groups,
e.g., experiment arms, with two possible plot types:
(1) Line plots of group-aggregated statistics over time, using the mean
    statistic (default) or any user-specified function and the
    \code{ggplot2::geom_line} function.
(2) Smoothed-mean plots of variables over time, using the
    \code{method = "gam"} argument in \code{ggplot2::geom_smooth}.
}
