% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{HPlot}
\alias{HPlot}
\title{Horizontal Bar or Point Plot}
\usage{
HPlot(
  data,
  x,
  y,
  type = c("bar", "point"),
  order_by = NULL,
  desc = TRUE,
  facet = NULL,
  facet_ncol = NULL,
  text = FALSE,
  text_nudge = 0,
  errorbar_vars = NULL,
  errorbar_height = 0.2,
  errorbar_size = 0.2,
  multi_fill = FALSE,
  vline = NULL,
  vline_type = 2,
  scale_log10 = FALSE,
  scale_percent = FALSE,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{A data frame containing data for the plot.}

\item{x}{Name of x-axis variable, which should be a numeric.}

\item{y}{Name of the y-axis variable, which should be categorical, i.e., a
character vector or a factor.}

\item{type}{The \code{ggplot2} geom to use, whether \code{"bar"} for a bar
plot, or \code{"point"} for a point plot.}

\item{order_by}{Name of numeric variable to order the y-axis variable by.}

\item{desc}{Whether to sort by descending order. Note that this specifies a
descending order from top of plot to bottom, which is in contrast to the
order that ggplot specifies, which is from bottom to top of plot.}

\item{facet}{Name of variable to facet by.}

\item{facet_ncol}{Number of columns for facetted plot.}

\item{text}{Whether to show the text label containing value of x-axis
variable.}

\item{text_nudge}{Number of points to nudge text label.}

\item{errorbar_vars}{A character vector containing names of the min and max
values of the errorbar. Vector must have at least two values if not NULL.}

\item{errorbar_height}{Height of errorbar.}

\item{errorbar_size}{Size of errorbar.}

\item{multi_fill}{Whether to fill bars with multiple colors or a single
color.}

\item{vline}{Line type for vertical lines.}

\item{scale_log10}{Whether to apply a log10 transformation to the x-axis.}

\item{scale_percent}{Whether format labels on the x-axis as percentages.}

\item{xlab}{Label for x-axis variable.}

\item{ylab}{Label for y-axis variable.}
}
\value{
A \code{ggplot2::ggplot} horizontal bar plot.
}
\description{
A generic wrapper function to create a horizontal bar or point plot using
\code{ggplot2::ggplot} with options for ordering the bars or points by a
specified variable. The function may be extended in the future to include
the option for horizontal boxplots.
}
