% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_samples.R
\name{MatchedData}
\alias{MatchedData}
\title{Matched Data Given MatchIt::matchit Object}
\usage{
MatchedData(matchit, data = NULL, match_var = "match")
}
\arguments{
\item{matchit}{A \code{MatchIt::matchit} object.}

\item{data}{An optional data frame passed into the \code{MatchIt::matchit}
function to create the \code{MatchIt::matchit} object. This is needed
if the \code{MatchIt::matchit} object does not contain any modeling data.}

\item{match_var}{Name of the variable to indicate matching.}
}
\value{
A data frame containing the matched sample.
}
\description{
Create from a \code{MatchIt::matchit} object a data frame containing the
matched sample, with treatment subjects in the first rows followed by their
matched control subjects, and a \code{match_var} variable to indicate the
matching.
}
