% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{MatchedPropDiffSE}
\alias{MatchedPropDiffSE}
\title{Standard Error of the Difference between Two Matched-Sample Proportions}
\usage{
MatchedPropDiffSE(b, c, n)
}
\arguments{
\item{b}{Count of observations with 1 for first sample and 0 for second.}

\item{c}{Count of observations with 0 for first sample and 1 for second.}

\item{n}{The sample size.}
}
\value{
Standard error estimate for the difference between the two
  matched-sample proportions.
}
\description{
Computes the standard error of the difference between the sample proportions
from two matched or paired samples. See Agresti and Min (2004).
}
