% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_samples.R
\name{MatchedSample}
\alias{MatchedSample}
\title{Matched Sample}
\usage{
MatchedSample(
  data,
  treat = "treat",
  covariates = 1,
  distance = NULL,
  estimand = c("att", "atu", "atc"),
  match = "match",
  ...
)
}
\arguments{
\item{data}{A data frame containing the treatment variable and covariates
for matching, or alternatively, a \code{psm} object returned by the
\code{FitPropensityScoreModel} function.}

\item{treat}{The name of the treatment variable.}

\item{covariates}{Character vector containing names of covariates.}

\item{distance}{An optional numeric vector of distance values, such that a
pairwise difference for any two observations gives the distance between
them. This could be a vector of propensity scores or any other measures to
derive pairwise distances.}

\item{estimand}{The causal estimand for the matching method to be applied
to the sample. Currently, only the average treatment effect on the treated
(ATT) population and the average treatment effect on the untreated or
control (ATU or ATC) population are available.}

\item{match}{Name of the variable to indicate matching.}

\item{...}{Additional arguments to be passed into the MatchIt::matchit
function.}
}
\value{
A data frame containing the treated sample and corresponding matches
  from the control sample.
}
\description{
Create a matched sample containing the treated observations and corresponding
matches from the control or untreated observations, with the following cases:
(1) Given the treatment variable and a measure to derive pairwise distances,
    e.g., a propensity score, and
(2) Given the treatment variable and a set of covariates.
}
