% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{MinDetectableEffectSize}
\alias{MinDetectableEffectSize}
\title{Minimum Detectable Effect Size}
\usage{
MinDetectableEffectSize(n1, n2 = NULL, sig_level = 0.05, power = 0.8)
}
\arguments{
\item{n1}{Size of the first sample, which is required.}

\item{n2}{Size for the second sample. If \code{NULL}, then it is
to be equal to the size \code{n1} of the first sample.}

\item{sig_level}{Significance level of the test.}

\item{power}{Power of the test.}
}
\value{
The minimum detectable effect size for the two-sample z test.
}
\description{
Computes the minimum detectable effect size for the power analysis of a
two-sample z test.
}
