% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{MultipleValuedVars}
\alias{MultipleValuedVars}
\title{Multiple-valued Variables}
\usage{
MultipleValuedVars(data, vars)
}
\arguments{
\item{data}{A data frame containing the variables specified in \code{vars}.}

\item{vars}{A character vector containing the string names of the variables.}
}
\value{
A character vector containing the names of only the multiple-valued
variables, i.e., variables with more than one unique value.
}
\description{
Given a vector of variable names in a data frame, return a vector with the
names of only the multiple-valued variables, i.e., variables with more than
one unique value.
}
