% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyApply}
\alias{MyApply}
\title{Customized Apply Function}
\usage{
MyApply(x, fun, ..., use_names = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{fun}{The function to apply.}

\item{...}{Additional arguments to pass to \code{base::sapply}.}

\item{use_names}{Value to pass into the \code{USE.NAMES} argument for
\code{base::sapply}.}
}
\value{
The result of applying the \code{base::sapply} function.
}
\description{
A wrapper function around \code{base::sapply} but with the \code{simplify}
argument set to FALSE instead of TRUE. Useful for converted a character
vector into a list with the vector as names.
}
