% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{OverlapPlot}
\alias{OverlapPlot}
\title{Overlap Plots}
\usage{
OverlapPlot(
  data,
  score = "score",
  treat = "treat",
  type = c("histogram", "density", "both"),
  bins = 50L,
  alpha = ifelse(type == "both", 0.35, 0.5),
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{A data frame containing the propensity score and the treatment
assignment variable, specified in, respectively, the \code{score} and
\code{treat} arguments.}

\item{score}{String name of the propensity score vector in \code{data}.}

\item{treat}{String name of the treatment assignment vector in \code{data}.}

\item{type}{Type of plot, either \code{"histogram"}, \code{"density"}, or
\code{"both"}.}

\item{bins}{Number of bins for the histogram. Defaults to 50.}

\item{alpha}{The \code{alpha} aesthetic for the \code{ggplot} object, to
determine opacity of the geom.}

\item{title}{Optional argument for the plot title.}

\item{xlab}{Optional argument for the x-axis label.}

\item{ylab}{Optional argument for the y-axis label.}
}
\value{
A \code{ggplot2::ggplot} object showing the overlap plot.
}
\description{
Create an "overlap" plot to evaluate the overlap of the propensity scores for
different study groups. The propensity score distributions are compared using
histograms, density plots, or both. This is essentially a wrapper around the
\code{personalized::check.overlap} function, which creates a
\code{ggplot2::ggplot} object. But the propensity score is passed directly
into the \code{personalized::check.overlap} function instead of having it
estimated by the function. Note the following for our implementation:
(1) The \code{type}, \code{bins}, and \code{alpha} arguments and their
    default values are the same as those in the
    \code{personalized::check.overlap} function.
(2) Since \code{personalized::check.overlap} function does not handle missing
    values, we impute missing values in the propensity score with the median
    value.
}
