% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{PropDiffSE}
\alias{PropDiffSE}
\title{Standard Error of the Difference between Two Independent-Sample Proportions}
\usage{
PropDiffSE(p1, p2, n1, n2)
}
\arguments{
\item{p1}{Estimate of the population proportion for the first sample.}

\item{p2}{Estimate of the population proportion for the second sample.}

\item{n1}{Size of the first sample.}

\item{n2}{Size of the second sample.}
}
\value{
Standard error estimate for the difference between the two
  independent-sample proportion.
}
\description{
Computes the standard error of the difference between the sample proportions
from two independent samples.
}
