% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_samples.R
\name{PropensityScoreWeight}
\alias{PropensityScoreWeight}
\title{Propensity Score Weights}
\usage{
PropensityScoreWeight(
  treat,
  score = NULL,
  weight = NULL,
  estimand = c("att", "ate", "atu", "atc", "ato"),
  normalized = FALSE
)
}
\arguments{
\item{treat}{Numeric vector with values 1 for treatment and 0 otherwise.}

\item{score}{The propensity score, i.e., the probability of treatment.}

\item{weight}{The PSWs passed in as an argument to infer the probability of
treatment.}

\item{estimand}{The causal estimand that determines the PSW, whether
\code{att} for the average treatment effect on the treated (ATT),
\code{ate} for the average treatment effect on the entire population (ATE),
\code{atu} or \code{atc} for the average treatment effect on the untreated
or control population (ATU or ATC), and \code{ato} for the average
treatment effect on the overlap population (ATO).}

\item{normalized}{Whether to normalize weights. Using normalized and
unnormalized weights correspond to using, respectively, the Hajek and
Horvitz-Thompson estimators. Applicable only for converting probabilities
to weights.}
}
\value{
A numeric vector with PSWs or the propensity scores.
}
\description{
Given a treatment indicator variable compute one of the following:
(1) Compute the propensity score weights (PSWs) given
    values for the probability of treatment, the propensity score.
(2) Compute the propensity score given the PSWs.
This is done for the following four kinds of causal estimands:
(1) The average treatment effect on the entire population (ATE),
(2) The average treatment effect on the treated (ATT),
(3) The average treatment effect on the untreated or control (ATU or ATC),
    and
(4) The average treatment effect on the overlap population (ATO).
Note that for the ATT, the probability of treatment can only be inferred for
the untreated or control population, while for the ATU/ATC, it can only be
inferred for the treated population.
}
\references{
\itemize{
\item Li, Fan, Kari Lock Morgan, and Alan M. Zaslavsky (2018). "Balancing
  Covariates via Propensity Score Weighting." \emph{Journal of the American
  Statistical Association}, 113(521): 390-400.
}
}
