% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{PropSE}
\alias{PropSE}
\title{Standard Error of a Sample Proportion}
\usage{
PropSE(p, n)
}
\arguments{
\item{p}{Estimate of the population proportion.}

\item{n}{The sample size.}
}
\value{
Standard error estimate for the sample binomial proportion.
}
\description{
Computes the standard error of the sample binomial proportion given an
}
