% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{RatioCI}
\alias{RatioCI}
\title{Confidence Interval for the Ratio of Means}
\usage{
RatioCI(
  method = c("fieller", "delta", "delta_log"),
  a,
  b,
  v11,
  v22,
  v12 = 0,
  s = 1,
  conf_level = 0.95,
  df = Inf,
  chg = FALSE
)
}
\arguments{
\item{method}{A string for the method name, either \code{"fieller"} for the
Fieller CI or \code{"delta"} for the delta method without log
transformation of the ratio.}

\item{a}{Sample estimate for mean in the numerator of the ratio.}

\item{b}{Sample estimate for mean in the denominator of the ratio.}

\item{v11}{Scaled variance of a.}

\item{v22}{Scaled variance of b.}

\item{v12}{Scaled covariance of a and b.}

\item{s}{The scaling factor for the covariance matrix for [a, b]. Defaults
to 1.}

\item{conf_level}{Confidence level for interval.}

\item{df}{Degree of freedom for t statistic for confidence interval.}

\item{chg}{Whether to convert the CI bounds from a ratio to a relative
change, i.e., ratio minus 1.}
}
\value{
A data frame containing the lower and upper bounds of the CI for the
  ratio of means.
}
\description{
Computes the confidence interval (CI) for a ratio of two means, given two
sample mean estimates and their covariance matrix, with the following
options:
(1) The Fieller CI, proposed in Fieller (1954).
(2) The delta method CI without log transformation of the ratio.
(3) The delta method CI with log transformation of the ratio.
See Hirschberg and Lye (2010) for a comparison of the first two approaches.
  Note that for the delta method without log transformation of the ratio,
the numerator and the denominator could be negative, as in the case for the
Fieller CI as well. But for the delta method with log transformation of the
ratio, both numerator and denominator need to be positive. The delta method
with log transformation of the ratio would not work for cases where the
numerator is zero, so it is less robust than the delta method without
log transformation for the ratio.
  The function is implemented in a "tidy" fashion by returning the results as
a data frame. The arguments \code{a, b, v11, v22, v12, s} can be numeric
vectors with multiple entries.
}
\references{
\itemize{
\item Fieller, E. C. (1954). "Some Problems in Interval Estimation."
   \emph{Journal of the Royal Statistical Society Series B,} 16: 175-185.
\item Hirschberg, Joe and Jenny Lye (2010). "A Geometric Comparison of the
   Delta and Fieller Confidence Intervals." \emph{The American Statistician,}
   64(3): 234-241.
}
}
