% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReplaceNA}
\alias{ReplaceNA}
\title{Replace NAs in Data Frame}
\usage{
ReplaceNA(data, vars = NULL, except = NULL, replace = 0)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{A character vector of names of variables to replace NAs for.}

\item{except}{A character vector of names of variables not to replace NAs
for.}

\item{replace}{A single character string for the replacement value, or
a named list of replacement strings. The names of the named list indicate
the variables to apply the replacement values to. For a named list
argument, \code{tidyr::replace_na} will be applied to the entire data frame
with this argument without any changes.}
}
\value{
A list of data frames split by the grouping variable(s).
}
\description{
Wrapper for the \code{tidyr::replace_na} function to make it easier for a
common use case, i.e., replace the NAs in most variables with a single value.
The highlight utility of this function is that we could apply the single
replacement value to all vectors of the same class as the replacement value,
e.g., all numeric vectors if the replacement value is \code{0}, all character
vectors if the replacement value is \code{"Missing"}, etc.
}
