% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Split}
\alias{Split}
\title{Split Data Frame}
\usage{
Split(data, split = NULL, drop_split = TRUE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{split}{A character vector containing the name(s) of the grouping
variable(s) for splitting.}

\item{drop_split}{Whether the grouping variable(s) should be dropped from
resulting list of data frames.}

\item{...}{Additional arguments to pass to \code{base::split}.}
}
\value{
A list of data frames split by the grouping variable(s).
}
\description{
Wrapper function around \code{base::split} to split data frames into a list
of sub data frames using one or more grouping variables.
}
