% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_subgroups.R
\name{SubgroupAnalysis}
\alias{SubgroupAnalysis}
\title{Subgroup Analysis}
\usage{
SubgroupAnalysis(
  data,
  y,
  model = NULL,
  vars = NULL,
  treat = NULL,
  estimand = c("ate", "att", "atc", "atu", "ato"),
  estimator = c("aipw", "tmle"),
  sep = "",
  label = "slice",
  include_all = TRUE,
  conf_level = 0.95,
  slice_name = "slice",
  percent_vars_resp = c("mean", "est", "lb", "ub", "se"),
  percent_vars_ate = c("mC", "mT", "diff", "eff", "lb", "ub", "se", "smd"),
  digits = digits
)
}
\arguments{
\item{data}{A data frame containing the response metric and the slicing
covariates, and optionally, the treatment variable, if the ATE is
estimated.}

\item{y}{Names of metric.}

\item{model}{An optional model object for predicting the metric for (1) or
estimating the ATE of the metric for (2). Currently, only the
\code{causal_forest} model object of the grf R package is supported.}

\item{vars}{A character vector containing the names of the slicing
covariates.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{estimand}{The causal estimand to be estimated by the causal model,
either \code{"ate"} for the average treatment effect on
the entire population (ATE), \code{"att"} for the average treatment effect
on the treated (ATT), or \code{"atu"} or \code{"atc"} for the average
treatment effect on the untreated or control (ATU or ATC), or \code{"ato"}
for the average treatment effect on the overlap population. Only applicable
when the \code{treat} argument is not NULL.}

\item{sep}{The separator character to separate the name of the slicing
variable and its slice value.}

\item{label}{The name of column in the output table for the slices.}

\item{include_all}{Whether to include the slice containing all the
observations, sometimes referred to as the global slice.}

\item{conf_level}{Confidence level for inference.}

\item{slice_name}{The name of the column containing the slice names.}

\item{percent_vars_resp}{The variables to percentize in the output table
if we are estimating the mean response.}

\item{percent_vars_ate}{The variables to percentize in the output table
if we are estimating the ATE.}

\item{digits}{Number of decimals for table statistics.}
}
\value{
A data frame containing the table of estimation statistics for each
  slice.
}
\description{
Perform a subgroup analysis for a single metric over different slices defined
by categorical covariates, with the following two cases:
(1) Estimate mean responses for each slice, if a treatment variable is not
    provided.
(2) Estimate average treatment effect (ATE) for each slice, if a treatment
    variable is provided.
}
\details{
Optionally takes the model-estimated predictions of individual responses or
individual treatment effects (ITEs) to perform the estimation of the mean
responses or ATEs for each slice. If a model is not provided, we provide
the sample mean for (1) and the sample difference between Treatment and
Control groups for (2). We will refer to subgroups as slices.
}
