% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_subgroups.R
\name{SubgroupAnalysisOverMetrics}
\alias{SubgroupAnalysisOverMetrics}
\title{Subgroup Analysis over Metrics}
\usage{
SubgroupAnalysisOverMetrics(
  data,
  y,
  models = NULL,
  treat = NULL,
  percent_vars_resp = c("mean", "est", "lb", "ub", "se"),
  percent_vars_ate = c("mC", "mT", "diff", "eff", "lb", "ub", "se", "smd"),
  digits = 2,
  ...
)
}
\arguments{
\item{data}{A data frame containing the response metric and the slicing
covariates, and optionally, the treatment variable, if the ATE is
estimated.}

\item{y}{A character vector of names of metric.}

\item{models}{Either a single model object or a list of model objects.}

\item{treat}{Name of treatment assignment indicator variable.}

\item{percent_vars_resp}{The variables to percentize in the output table
if we are estimating the mean response.}

\item{percent_vars_ate}{The variables to percentize in the output table
if we are estimating the ATE.}

\item{digits}{Number of decimals for table statistics.}
}
\value{
A data frame containing table of estimation statistics for each
  slice if there is only one metric, and a list of such data frames if
  there is more than one metric.
}
\description{
A wrapper function to run the \code{SubgroupAnalysis} function for multiple
metrics. See documentation for the \code{SubgroupAnalysis} function for
further details.
}
