% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{TrimLinearModelFit}
\alias{TrimLinearModelFit}
\title{Trim Linear Model Fit}
\usage{
TrimLinearModelFit(m)
}
\arguments{
\item{m}{A model fit object generated by \code{stats::lm},
\code{stats::glm}, or \code{survey::svyglm}.}
}
\value{
A trimmed version of the model fit object.
}
\description{
Remove model data from the linear model fit generated by \code{stats::lm},
\code{stats::glm}, or \code{survey::svyglm}:
(1) For \code{stats::lm} and \code{stats::glm}, the \code{model} object is
    removed.
(2) For \code{survey::svyglm}, the \code{model} and \code{data} objects are
    removed.
}
