% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_samples.R
\name{UnadjustedSample}
\alias{UnadjustedSample}
\title{Unadjusted Sample}
\usage{
UnadjustedSample(
  data,
  treat = "treat",
  outcomes = NULL,
  covariates = NULL,
  ids = NULL
)
}
\arguments{
\item{data}{A data frame containing the treatment variable.}

\item{treat}{The name of the treatment variable.}

\item{outcomes}{Character vector containing names of outcomes.}

\item{covariates}{Character vector containing names of covariates.}

\item{ids}{Character vector containing names of ids.}
}
\value{
A data frame containing the original data with rows having missing
  values for treatment removed.
}
\description{
Create an unadjusted sample for analysis, namely, by including only the
columns for treatment, outcomes, and covariates and removing rows where
the treatment variable is missing.
}
