% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{UniqueValueCount}
\alias{UniqueValueCount}
\title{Unique Value Count}
\usage{
UniqueValueCount(data, vars)
}
\arguments{
\item{data}{A data frame containing the variables specified in \code{vars}.}

\item{vars}{A character vector containing the string names of the variables.}
}
\value{
A named numeric vector containing the number of unique values for
  variable, with the variable names in \code{vars} as the entry names.
}
\description{
Given a vector of column variables in a data frame, return a vector of the
number of unique values for each variable.
}
