% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Unsplit}
\alias{Unsplit}
\title{Unsplit a List of Data Frames}
\usage{
Unsplit(data, label = "label")
}
\arguments{
\item{data}{A data frame.}

\item{label}{The name of the first column added to data frame unsplit from
the list.}

\item{drop_split}{Whether the grouping variable(s) should be dropped from
resulting list of data frames.}

\item{...}{Additional arguments to pass to \code{base::split}.}
}
\value{
A list of data frames split by the grouping variable(s).
}
\description{
Combines a named list of similar data frames into a single data frame where
the first column contains the list names. Function is so named, because it is
the reverse operation of the \code{Split} function. It functions similarly to
the \code{base::unsplit} function but does not use it. Data frames are
converted to \code{data.table} objects for performance.
}
