% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_samples.R
\name{WeightedData}
\alias{WeightedData}
\title{Weighted Data Sample [DEPRECATED: Don't Use.]}
\usage{
WeightedData(psm, weights = "weights", type = c("att", "ate"))
}
\arguments{
\item{psm}{An object of class \code{psm}, which could be created by the
\code{FitPropensityScoreModel} function.}

\item{weights}{Name of weight variable.}

\item{type}{Type of the IPTW, whether ATT or ATE.}
}
\value{
A data frame containing the original model data with the IPTW added
  to it.
}
\description{
Given a \code{psm} object, return a data frame with
inverse-probability-of-treatment weights (IPTW) added to it. The type of
IPTW can be specified for either the average treatment effect on the treated
(ATT) or the average treatment effect (ATE).
}
