% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_effects.R
\name{WeightedEffectTable}
\alias{WeightedEffectTable}
\title{Weighted Effect Table}
\usage{
WeightedEffectTable(
  data,
  y,
  weights = "weights",
  use_svyglm = FALSE,
  hc = c("hc1", "hc0", "hc2", "hc3", "hc4", "none"),
  ...
)
}
\arguments{
\item{data}{A data frame containing metrics for treatment and control
subjects, an indicator variable for treatment assignment and optionally
additional covariates and an integer variable for matching of pairs.}

\item{y}{Character vector containing names of variables to estimate effects
for.}

\item{weights}{Name of variable containing inverse-probability-of-treatment
weights (IPTW) for a weighted data sample. We recommend using a
heteroskedasticity-corrected standard errors in this case, e.g.,
\code{hc = "hc0"} instead of \code{hc = "none"}.}

\item{use_svyglm}{Whether to use \code{survey::svyglm} to estimate an IPTW
treatment effect effect with its standard errors. The default is FALSE,
in which case \code{stats::lm} is used.}

\item{hc}{Option for computing heteroskedasticity-corrected SEs using
\code{car::hccm}. Defaults to \code{"hc1"}. The other values are
\code{"hc0"}, \code{"hc2"}, \code{"hc3"}, \code{"hc4"}, and \code{"none"}.
See documentation for \code{car::hccm} for details.}

\item{...}{Other arguments passed into \code{AdjustedEffectTable}, not
included above and not one of the following: \code{include_pre},
\code{interact_pre}, and \code{post_suffix}, which are already have set
values passed into the function. See documentation for
\code{AdjustedEffectTable} for description of other arguments.}
}
\value{
A data frame containing the table of unadjusted effects for the
  IPTW sample with relevant statistics.
}
\description{
Wrapper for \code{AdjustedEffectTable} to estimate unadjusted effects from
inverse-probability-of-treatment-weighted (IPTW) samples, using either
\code{stats::lm} with Huber-White heteroskedasticity-corrected (HC) standard
errors or \code{survey::svyglm}, which also implements the HC errors. From
examples we have seen, we find that the \code{hc = "hc1"} option gives SE
estimates for \code{stats::lm} closest to those obtained using
\code{survey::svyglm}, although the other HC options tend to give very
similar results as well.
}
