% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_samples.R
\name{WeightedSample}
\alias{WeightedSample}
\title{Weighted Sample}
\usage{
WeightedSample(
  data,
  treat = "treat",
  covariates = NULL,
  estimand = c("att", "ate", "atu", "atc", "ato"),
  score = "score",
  weights = "weights",
  type = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing the treatment variable and covariates
for estimating the weights for the sample, or alternatively, a \code{psm}
object returned by the \code{FitPropensityScoreModel} function.}

\item{treat}{The name of the treatment variable.}

\item{covariates}{Character vector containing names of covariates.}

\item{estimand}{The causal estimand for the propensity score weight or
balancing weight to be applied to the sample.}

\item{score}{Name of the variable containing the estimated propensity score.}

\item{weights}{Name of weight variable.}

\item{type}{Deprecated name for the \code{estimand} argument. Kept here
for backward compatibility.}

\item{...}{Additional arguments to be passed into the
\code{FitPropensityScoreModel} function.}
}
\value{
A data frame containing the original model data with the propensity
  score and propensity score weights or balancing weights added to it.
}
\description{
Given the names of the treatment variable and covariates, return a data frame
with the propensity score and propensity score weights (PSWs) or balancing
weights attached to it. Relies on the \code{FitPropensityScoreModel} function
to compute the propensity score and the weights. Can also take a \code{psm}
object returned by the \code{FitPropensityScoreModel} function in the
\code{data} argument and extracts the data, score, and weights from the
the object. See the documentation for \code{FitPropensityScoreModel} for
details on the methods used to estimate the propensity score and weights.
}
